Refactor docs and dashboard layouts into one
