+++ apps/web/src/app/(docs)/docs/layout.tsx
@@ -1,15 +1,8 @@
-import Script from 'next/script'
-import { Metadata } from 'next'
 import glob from 'fast-glob'
-import { Analytics } from '@vercel/analytics/react'
 
-import { Providers } from '@/app/providers'
 import { Layout } from '@/components/Layout'
 
-import '@/styles/tailwind.css'
-import { PostHogAnalytics } from '@/utils/usePostHog'
 import { Section } from '@/components/SectionProvider'
-import Canonical from '@/components/Navigation/canonical'
 
 declare global {
   // eslint-disable-next-line @typescript-eslint/no-namespace
@@ -20,25 +13,8 @@ declare global {
   }
 }
 
-export const metadata: Metadata = {
-  // TODO: Add metadataBase
-  // metadataBase: ''
-  title: {
-    template: '%s - E2B Docs',
-    default: 'E2B Docs - Code Interpreting for AI apps',
-  },
-  description: 'Open-source secure sandboxes for AI code execution',
-  twitter: {
-    title: 'E2B Docs - Code Interpreting for AI apps',
-    description: 'Open-source secure sandboxes for AI code execution',
-  },
-  openGraph: {
-    title: 'E2B Docs - Code Interpreting for AI apps',
-    description: 'Open-source secure sandboxes for AI code execution',
-  },
-}
 
-export default async function RootLayout({ children }) {
+export default async function DocsLayout({ children }) {
   const pages = await glob('**/*.mdx', { cwd: 'src/app/(docs)/docs' })
   const allSectionsEntries = (await Promise.all(
     pages.map(async filename => [
@@ -47,30 +23,10 @@ export default async function RootLayout({ children }) {
     ]),
   )) as Array<[string, Array<Section>]>
   const allSections = Object.fromEntries(allSectionsEntries)
-    
-  return (
-    <html
-      lang="en"
-      className="h-full"
-      suppressHydrationWarning
-    >
-      <head>
-        <Canonical />
-      </head>
-      <body className="flex min-h-full bg-white antialiased dark:bg-zinc-900">
-        <Providers>
-          <div className="w-full">
-            <Layout allSections={allSections}>
-              {children}
-              <PostHogAnalytics />
-              <Analytics />
-            </Layout>
-          </div>
-        </Providers>
 
-        <Script src="https://js.chatlio.com/widget.js" strategy="lazyOnload" />
-        <chatlio-widget widgetid={process.env.NEXT_PUBLIC_CHATLIO_WIDGET_ID} disable-favicon-badge></chatlio-widget>
-      </body>
-    </html>
+  return (
+      <Layout allSections={allSections}>
+        {children}
+      </Layout>
   )
 }
